api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn[standard]==0.27.1",
    "python-dotenv==1.0.1",
    "llama-index>=0.10.68",
    "llama-index-core>=0.10.68.post1",
    "llama-index-storage-docstore-postgres>=0.1.3",
    "llama-index-storage-kvstore-postgres>=0.1.2",
    "llama-index-storage-index-store-postgres>=0.2.0",
    "llama-index-vector-stores-postgres>=0.1.14",
    "llama-index-llms-openai>=0.3.25",
    "fastapi>=0.115.11",
    "pydantic>=2.10.6",
    "markitdown[all]~=0.1.0a1",
    "asyncpg>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.black]
line-length = 88
target-version = ["py312"]

[dependency-groups]
dev = [
    "bpython>=0.25",
    "ruff>=0.9.10",
]
